requirements.txt
annotated-types==0.7.0
google-auth
anyio==4.6.2.post1
asyncpg==0.30.0
build==1.2.2.post1
CacheControl==0.14.1
certifi==2024.8.30
charset-normalizer==3.4.0
cleo==2.1.0
click==8.1.7
colorama==0.4.6
crashtest==0.4.1
databases==0.9.0
distlib==0.3.9
dnspython==2.7.0
dulwich==0.21.7
email_validator==2.2.0
fastapi==0.115.5
fastapi-cli==0.0.5
fastjsonschema==2.20.0
filelock==3.16.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
idna==3.10
installer==0.7.0
jaraco.classes==3.4.0
Jinja2==3.1.4
keyring==24.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.5.0
msgpack==1.1.0
packaging==24.2
pexpect==4.9.0
pkginfo==1.11.2
platformdirs==4.3.6
poetry==1.8.4
poetry-core==1.9.1
poetry-plugin-export==1.8.0
ptyprocess==0.7.0
pydantic==2.10.2
pydantic_core==2.27.1
Pygments==2.18.0
pyproject_hooks==1.2.0
python-dotenv==1.0.1
python-multipart==0.0.17
pywin32-ctypes==0.2.3
PyYAML==6.0.2
RapidFuzz==3.10.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlmodel==0.0.22
starlette==0.41.3
tomlkit==0.13.2
trove-classifiers==2024.10.21.16
typer==0.13.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
virtualenv==20.28.0
watchfiles==1.0.0
websockets==14.1
python-jose
pydantic_settings
passlib
psycopg2-binary
pytesseract
PyPDF2
pymongo
openai
PyMuPDF
motor
pdf2image
pyjwt
bcrypt
google_auth_oauthlib
google-api-python-client
